+++ src/fides/api/models/privacy_request.py
@@ -1307,6 +1307,7 @@ class ProvidedIdentityType(EnumType):
     ga_client_id = "ga_client_id"
     ljt_readerID = "ljt_readerID"
     fides_user_device_id = "fides_user_device_id"
+    external_id = "external_id"
 
 
 class ProvidedIdentity(Base):  # pylint: disable=R0904
@@ -1495,6 +1496,11 @@ class Consent(Base):
 class ConsentRequest(IdentityVerificationMixin, Base):
     """Tracks consent requests."""
 
+    property_id = Column(
+        String,
+        index=True,
+        nullable=True,
+    )
     provided_identity_id = Column(
         String, ForeignKey(ProvidedIdentity.id), nullable=False
     )